First shot working with simple array.
If I recall correctly, there's two sources this stuff was coming from -- one
all sort of hardcoded up in a specific app, and the bulk of that is still in
the main Batch class ... but then other sources had already started breaking
things out to the separate source classes. So, it's a bit of a hodge-podge, and
then was never really designed to just go for it with plain arrays.