Version 0.1.3
Bug fix ActiveRecordSource after changes in 0.1.2. 0.1.2 was short term changes
to get some basic Array support working, but to do that I had to move a lot of
stuff out of Batch into HashRowsSource ... but ActiveRecordSource was also
needing a lot of the same thing. ActiveRecordSource could almost simply mixin
HashRowsSource, except for the text_table method itself, thus the introduction
of a HashRowsBase module that can be an intermediary.